+++ PolyLib/gfx/guiButtons.hpp
@@ -6,7 +6,7 @@
 // GUIHeader Box for Panel Selection
 class Header_PanelBox {
   public:
-    Header_PanelBox(uint16_t x, uint16_t y, uint16_t width, uint16_t heigth) {
+    Header_PanelBox(uint32_t x, uint32_t y, uint32_t width, uint32_t heigth) {
         this->x = x;
         this->y = y;
         this->width = width;
@@ -18,16 +18,16 @@ class Header_PanelBox {
   private:
     const GUI_FONTINFO *font = &GUI_FontBahnschrift24_FontInfo;
 
-    uint16_t x;
-    uint16_t y;
-    uint16_t width;
-    uint16_t heigth;
+    uint32_t x;
+    uint32_t y;
+    uint32_t width;
+    uint32_t heigth;
     uint8_t selected = 0;
 };
 class Footer_PanelBox {
   public:
-    Footer_PanelBox(actionHandle *mainFunction, actionHandle *secondaryFunction, uint16_t x, uint16_t y, uint16_t width,
-                    uint16_t heigth) {
+    Footer_PanelBox(actionHandle *mainFunction, actionHandle *secondaryFunction, uint32_t x, uint32_t y, uint32_t width,
+                    uint32_t heigth) {
         this->main = mainFunction;
         this->secondary = secondaryFunction;
 
@@ -45,16 +45,16 @@ class Footer_PanelBox {
 
     actionHandle *main;
     actionHandle *secondary;
-    uint16_t x;
-    uint16_t y;
-    uint16_t width;
-    uint16_t heigth;
-    uint16_t fontShift = 0;
+    uint32_t x;
+    uint32_t y;
+    uint32_t width;
+    uint32_t heigth;
+    uint32_t fontShift = 0;
 };
 
 class Side_PanelBox {
   public:
-    Side_PanelBox(ButtonActionHandle *actionHandle, uint16_t x, uint16_t y, uint16_t width, uint16_t heigth) {
+    Side_PanelBox(ButtonActionHandle *actionHandle, uint32_t x, uint32_t y, uint32_t width, uint32_t heigth) {
         this->actionHandle = actionHandle;
 
         this->x = x;
@@ -70,17 +70,17 @@ class Side_PanelBox {
     const GUI_FONTINFO *fontSmall = &GUI_FontBahnschrift24_FontInfo;
 
     ButtonActionHandle *actionHandle;
-    uint16_t x;
-    uint16_t y;
-    uint16_t width;
-    uint16_t heigth;
+    uint32_t x;
+    uint32_t y;
+    uint32_t width;
+    uint32_t heigth;
 };
 
 // Header
 class GUIHeader {
   public:
-    void init(std::vector<GUIPanelBase *> *panels, uint8_t *activePanelID, uint16_t width, uint16_t height,
-              uint16_t x = 0, uint16_t y = 0);
+    void init(std::vector<GUIPanelBase *> *panels, uint8_t *activePanelID, uint32_t width, uint32_t height,
+              uint32_t x = 0, uint32_t y = 0);
     void Draw();
 
   private:
@@ -88,43 +88,43 @@ class GUIHeader {
     uint8_t *activePanelID = nullptr;
     std::vector<GUIPanelBase *> *panels;
     std::vector<Header_PanelBox> boxes;
-    uint16_t panelWidth;
-    uint16_t panelHeight;
-    uint16_t panelCount;
-    uint16_t panelAbsX;
-    uint16_t panelAbsY;
+    uint32_t panelWidth;
+    uint32_t panelHeight;
+    uint32_t panelCount;
+    uint32_t panelAbsX;
+    uint32_t panelAbsY;
 };
 
 // FOOTER
 class GUIFooter {
   public:
-    void init(uint16_t width, uint16_t height, uint16_t x = 0, uint16_t y = LCDHEIGHT);
+    void init(uint32_t width, uint32_t height, uint32_t x = 0, uint32_t y = LCDHEIGHT);
     void Draw();
 
   private:
     // Boxes
     std::vector<Footer_PanelBox> boxes;
 
-    uint16_t panelWidth;
-    uint16_t panelHeight;
-    uint16_t panelCount = 4;
-    uint16_t panelAbsX;
-    uint16_t panelAbsY;
+    uint32_t panelWidth;
+    uint32_t panelHeight;
+    uint32_t panelCount = 4;
+    uint32_t panelAbsX;
+    uint32_t panelAbsY;
 };
 
 // SIDE
 class GUISide {
   public:
-    void init(uint16_t width = BOARDERWIDTH, uint16_t height = LCDHEIGHT, uint16_t y = 0, uint16_t x = 0);
+    void init(uint32_t width = BOARDERWIDTH, uint32_t height = LCDHEIGHT, uint32_t y = 0, uint32_t x = 0);
     void Draw();
 
   private:
     // Boxes
     std::vector<Side_PanelBox> boxes;
 
-    uint16_t panelWidth;
-    uint16_t panelHeight;
-    uint16_t panelCount = 3;
-    uint16_t panelAbsX;
-    uint16_t panelAbsY;
+    uint32_t panelWidth;
+    uint32_t panelHeight;
+    uint32_t panelCount = 3;
+    uint32_t panelAbsX;
+    uint32_t panelAbsY;
 };
