+++ PolyLib/gfx/guiPanelEffect.hpp
@@ -8,7 +8,7 @@ typedef enum { WAVE, PHASE, CRUSH } effectSelect;
 
 class GUIPanelEffect : public GUIPanelBase {
   public:
-    void init(uint16_t width, uint16_t height, uint16_t x = 0, uint16_t y = 0, uint8_t pathVisible = 1);
+    void init(uint32_t width, uint32_t height, uint32_t x = 0, uint32_t y = 0, uint8_t pathVisible = 1);
     void Draw();
 
     void updateEntrys(BaseModule *module);
