+++ PolyLib/gfx/guiPanelPath.cpp
@@ -1,7 +1,7 @@
 #ifdef POLYCONTROL
 #include "guiPanelPath.hpp"
 
-void GUIPanelPath::init(uint16_t width, uint16_t height, uint16_t x, uint16_t y) {
+void GUIPanelPath::init(uint32_t width, uint32_t height, uint32_t x, uint32_t y) {
     panelWidth = width;
     panelHeight = height;
     panelAbsX = x;
