+++ PolyLib/gfx/guiPanelPath.hpp
@@ -3,7 +3,7 @@
 // PANEL Path Type
 class GUIPanelPath {
   public:
-    void init(uint16_t width, uint16_t height, uint16_t x = 0, uint16_t y = 0);
+    void init(uint32_t width, uint32_t height, uint32_t x = 0, uint32_t y = 0);
 
     void Draw(uint8_t onlyLayer = 0);
