+++ PolyLib/gfx/guiPanelPreset.hpp
@@ -4,7 +4,7 @@
 
 class GUIPanelPreset : public GUIPanelBase {
   public:
-    void init(uint16_t width, uint16_t height, uint16_t x = 0, uint16_t y = 0, std::string name = "", uint8_t id = 0,
+    void init(uint32_t width, uint32_t height, uint32_t x = 0, uint32_t y = 0, std::string name = "", uint8_t id = 0,
               uint8_t pathVisible = 1);
     void Draw();
