+++ PolyLib/gfx/guiPanelStart.cpp
@@ -2,7 +2,7 @@
 
 #include "guiPanelStart.hpp"
 
-void GUIPanelStart::init(uint16_t width, uint16_t height, uint16_t x, uint16_t y) {
+void GUIPanelStart::init(uint32_t width, uint32_t height, uint32_t x, uint32_t y) {
     panelWidth = width;
     panelHeight = height;
     panelAbsX = x;
