+++ PolyLib/gfx/guiPanelString.cpp
@@ -7,7 +7,7 @@
 extern devManager deviceManager;
 extern CircularBuffer<char, 1024> consoleBuffer;
 
-void GUIPanelString::init(uint16_t width, uint16_t height, uint16_t x, uint16_t y) {
+void GUIPanelString::init(uint32_t width, uint32_t height, uint32_t x, uint32_t y) {
     panelWidth = width;
     panelHeight = height;
     panelAbsX = x;
