+++ PolyLib/gfx/guiPanelVoice.hpp
@@ -5,7 +5,7 @@
 
 class GUIPanelVoice : public GUIPanelBase {
   public:
-    void init(uint8_t layerID, uint16_t width, uint16_t height, uint16_t x = 0, uint16_t y = 0);
+    void init(uint8_t layerID, uint32_t width, uint32_t height, uint32_t x = 0, uint32_t y = 0);
     void Draw();
 
   private:
